requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==2.5.1
pymysql==1.0.3
boto3==1.28.13
requests==2.31.0
python-decouple==3.8